Break mapper core out into separate module. Improve debug logging in mapper. Add new explain mbid mapping labs query
