+++ listenbrainz/labs_api/labs/api/recording_search.py
@@ -7,7 +7,7 @@ from datasethoster.main import register_query
 from unidecode import unidecode
 
 from listenbrainz import config
-from listenbrainz.labs_api.labs.api.mbid_mapping import prepare_query, COLLECTION_NAME
+from listenbrainz.mbid_mapping_writer.mbid_mapper import prepare_query, COLLECTION_NAME
 
 
 NUM_TYPOS = 5
