+++ listenbrainz/labs_api/labs/main.py
@@ -6,6 +6,7 @@ from listenbrainz.labs_api.labs.api.artist_credit_from_artist_mbid import Artist
 from listenbrainz.labs_api.labs.api.artist_credit_from_artist_msid import ArtistCreditIdFromArtistMSIDQuery
 from listenbrainz.labs_api.labs.api.recording_from_recording_mbid import RecordingFromRecordingMBIDQuery
 from listenbrainz.labs_api.labs.api.mbid_mapping import MBIDMappingQuery
+from listenbrainz.labs_api.labs.api.explain_mbid_mapping import ExplainMBIDMappingQuery
 from listenbrainz.labs_api.labs.api.year_from_artist_credit_recording import YearFromArtistCreditRecordingQuery
 from listenbrainz.labs_api.labs.api.recording_search import RecordingSearchQuery
 from listenbrainz.labs_api.labs.api.artist_credit_recording_lookup import ArtistCreditRecordingLookupQuery
@@ -16,6 +17,7 @@ register_query(ArtistCreditIdFromArtistMBIDQuery())
 register_query(ArtistCreditIdFromArtistMSIDQuery())
 register_query(RecordingFromRecordingMBIDQuery())
 register_query(MBIDMappingQuery())
+register_query(ExplainMBIDMappingQuery())
 register_query(YearFromArtistCreditRecordingQuery())
 register_query(RecordingSearchQuery())
 register_query(ArtistCreditRecordingLookupQuery())
